Rename and fix sync visualizer behaviour
