+++ Runtime/Input/Controllers/IControllerVisualizer.cs
@@ -19,11 +19,14 @@ namespace RealityToolkit.Input.Controllers
         GameObject GameObject { get; }
 
         /// <summary>
-        /// If set, the <see cref="IControllerVisualizer"/>'s <see cref="GameObject"/>
-        /// pose in the scene will be synced with the assigned <see cref="Transform"/>
-        /// instead of synchronizing with the actual source pose data.
+        /// If set, the <see cref="IControllerPoseSynchronizer.PoseDriver"/>'s pose in the scene
+        /// is override and the actual <see cref="IController.InputSource"/> pose is ignored.
         /// </summary>
-        Transform VisualizerPoseOverrideSource { get; set; }
+        /// <remarks>
+        /// When active, set the <see cref="IControllerPoseSynchronizer.PoseDriver"/> position and rotation
+        /// directly to move the <see cref="IControllerVisualizer"/>.
+        /// </remarks>
+        bool OverrideSourcePose { get; set; }
 
         /// <summary>
         /// The <see cref="Transform"/> defining where <see cref="Interactors.IPokeInteractor"/>s
